Refactor MicroProfile Metrics CDI 1.x bundle to remove MetricsConfiguration, MetricsConfigurationEvent, MetricNameFactory, MetricParameters and consolidate MetricName and SeMetricName into a singular MetricName plus shuffle packages around
