+++ dev/com.ibm.ws.microprofile.metrics.cdi/src/com/ibm/ws/microprofile/metrics/cdi/interceptors/MeteredInterceptor.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
+ * Copyright (c) 2017, 2019 IBM Corporation and others.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -21,7 +21,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package io.astefanutti.metrics.cdi;
+package com.ibm.ws.microprofile.metrics.cdi.interceptors;
 
 import java.lang.reflect.AnnotatedElement;
 import java.lang.reflect.Member;
@@ -36,14 +36,16 @@ import javax.interceptor.AroundTimeout;
 import javax.interceptor.Interceptor;
 import javax.interceptor.InvocationContext;
 
+import org.eclipse.microprofile.metrics.Meter;
 import org.eclipse.microprofile.metrics.MetricRegistry;
-import org.eclipse.microprofile.metrics.Timer;
-import org.eclipse.microprofile.metrics.annotation.Timed;
+import org.eclipse.microprofile.metrics.annotation.Metered;
 
-@Timed
+import io.astefanutti.metrics.cdi.MetricResolver;
+
+@Metered
 @Interceptor
 @Priority(Interceptor.Priority.LIBRARY_BEFORE + 10)
-/* package-private */ class TimedInterceptor {
+public class MeteredInterceptor {
 
     private final Bean<?> bean;
 
@@ -52,38 +54,34 @@ import org.eclipse.microprofile.metrics.annotation.Timed;
     private final MetricResolver resolver;
 
     @Inject
-    private TimedInterceptor(@Intercepted Bean<?> bean, MetricRegistry registry, MetricResolver resolver) {
+    private MeteredInterceptor(@Intercepted Bean<?> bean, MetricRegistry registry, MetricResolver resolver) {
         this.bean = bean;
         this.registry = registry;
         this.resolver = resolver;
     }
 
     @AroundConstruct
-    private Object timedConstructor(InvocationContext context) throws Exception {
-        return timedCallable(context, context.getConstructor());
+    private Object meteredConstructor(InvocationContext context) throws Exception {
+        return meteredCallable(context, context.getConstructor());
     }
 
     @AroundInvoke
-    private Object timedMethod(InvocationContext context) throws Exception {
-        return timedCallable(context, context.getMethod());
+    private Object meteredMethod(InvocationContext context) throws Exception {
+        return meteredCallable(context, context.getMethod());
     }
 
     @AroundTimeout
-    private Object timedTimeout(InvocationContext context) throws Exception {
-        return timedCallable(context, context.getMethod());
+    private Object meteredTimeout(InvocationContext context) throws Exception {
+        return meteredCallable(context, context.getMethod());
     }
 
-    private <E extends Member & AnnotatedElement> Object timedCallable(InvocationContext context, E element) throws Exception {
-        String name = resolver.timed(bean.getBeanClass(), element).metricName();
-        Timer timer = (Timer) registry.getMetrics().get(name);
-        if (timer == null)
-            throw new IllegalStateException("No timer with name [" + name + "] found in registry [" + registry + "]");
+    private <E extends Member & AnnotatedElement> Object meteredCallable(InvocationContext context, E element) throws Exception {
+        String name = resolver.metered(bean.getBeanClass(), element).metricName();
+        Meter meter = (Meter) registry.getMetrics().get(name);
+        if (meter == null)
+            throw new IllegalStateException("No meter with name [" + name + "] found in registry [" + registry + "]");
 
-        Timer.Context time = timer.time();
-        try {
-            return context.proceed();
-        } finally {
-            time.stop();
-        }
+        meter.mark();
+        return context.proceed();
     }
 }
