+++ dev/com.ibm.ws.microprofile.metrics.cdi/src/io/astefanutti/metrics/cdi/MetricProducer.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
+ * Copyright (c) 2017, 2019 IBM Corporation and others.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -41,7 +41,7 @@ import org.eclipse.microprofile.metrics.Timer;
 @Alternative
 @Dependent
 @Priority(Interceptor.Priority.LIBRARY_BEFORE)
-/* package-private */ final class MetricProducer {
+public final class MetricProducer {
 
     @Produces
     private static Counter counter(InjectionPoint ip, MetricRegistry registry, MetricName metricName, MetricsExtension extension) {
